+++ builtin/v11/market/cbor_gen.go
@@ -543,7 +543,7 @@ func (t *PublishStorageDealsParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Deals ([]market.ClientDealProposal) (slice)
-	if len(t.Deals) > cbg.MaxLength {
+	if uint64(len(t.Deals)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Deals was too long")
 	}
 
@@ -638,7 +638,7 @@ func (t *PublishStorageDealsReturn) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.IDs ([]abi.DealID) (slice)
-	if len(t.IDs) > cbg.MaxLength {
+	if uint64(len(t.IDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.IDs was too long")
 	}
 
@@ -753,7 +753,7 @@ func (t *ActivateDealsParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.DealIDs ([]abi.DealID) (slice)
-	if len(t.DealIDs) > cbg.MaxLength {
+	if uint64(len(t.DealIDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.DealIDs was too long")
 	}
 
@@ -778,6 +778,7 @@ func (t *ActivateDealsParams) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -895,7 +896,7 @@ func (t *ActivateDealsResult) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.VerifiedInfos ([]market.VerifiedDealInfo) (slice)
-	if len(t.VerifiedInfos) > cbg.MaxLength {
+	if uint64(len(t.VerifiedInfos)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.VerifiedInfos was too long")
 	}
 
@@ -999,7 +1000,7 @@ func (t *VerifyDealsForActivationParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Sectors ([]market.SectorDeals) (slice)
-	if len(t.Sectors) > cbg.MaxLength {
+	if uint64(len(t.Sectors)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Sectors was too long")
 	}
 
@@ -1094,7 +1095,7 @@ func (t *VerifyDealsForActivationReturn) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Sectors ([]market.SectorDealData) (slice)
-	if len(t.Sectors) > cbg.MaxLength {
+	if uint64(len(t.Sectors)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Sectors was too long")
 	}
 
@@ -1189,7 +1190,7 @@ func (t *ComputeDataCommitmentParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Inputs ([]*market.SectorDataSpec) (slice)
-	if len(t.Inputs) > cbg.MaxLength {
+	if uint64(len(t.Inputs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Inputs was too long")
 	}
 
@@ -1294,7 +1295,7 @@ func (t *ComputeDataCommitmentReturn) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.CommDs ([]typegen.CborCid) (slice)
-	if len(t.CommDs) > cbg.MaxLength {
+	if uint64(len(t.CommDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.CommDs was too long")
 	}
 
@@ -1560,6 +1561,7 @@ func (t *GetDealTermReturn) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -1674,6 +1676,7 @@ func (t *GetDealActivationReturn) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -1779,7 +1782,7 @@ func (t *OnMinerSectorsTerminateParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.DealIDs ([]abi.DealID) (slice)
-	if len(t.DealIDs) > cbg.MaxLength {
+	if uint64(len(t.DealIDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.DealIDs was too long")
 	}
 
@@ -2255,7 +2258,7 @@ func (t *SectorDeals) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.DealIDs ([]abi.DealID) (slice)
-	if len(t.DealIDs) > cbg.MaxLength {
+	if uint64(len(t.DealIDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.DealIDs was too long")
 	}
 
@@ -2553,7 +2556,7 @@ func (t *SectorDataSpec) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.DealIDs ([]abi.DealID) (slice)
-	if len(t.DealIDs) > cbg.MaxLength {
+	if uint64(len(t.DealIDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.DealIDs was too long")
 	}
 
@@ -2578,6 +2581,7 @@ func (t *SectorDataSpec) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
