+++ builtin/v9/init/init_actor_state.go
@@ -130,7 +130,7 @@ func (d *InstalledActors) MarshalCBOR(w io.Writer) error {
 
 	scratch := make([]byte, 9)
 
-	if len(d.Entries) > cbg.MaxLength {
+	if uint64(len(d.Entries)) > cbg.MaxLength {
 		return fmt.Errorf("too many installed actor entries")
 	}
